Add files via upload
Scripts for retrieving LAM files from archive, coarse-graining, calculating advective increments. Also scripts for retrieving global model data, extracting from it the data for the location of a LAM and then catenating all the global model data together into along curtain. Script for plotting location of LAM domains as a normal map and as columns sticking out from a sphere. Also code for adding/correcting the meta data in the nc to make them suitable for uploading to a data repository.